+++ src/hooks/Event/useFetchAllEvent.ts
@@ -1,6 +1,7 @@
 import { useQuery, UseQueryResult } from "@tanstack/react-query";
 import EventService from "@/services/Event";
-import { TAdministrationResponse } from "@/services/Administration/types";
+// import { TAdministrationResponse } from "@/services/Administration/types";
+import { TAdministrationResponse } from "@/modules/Administration/type";
 
 
 export const useFetchAllEvent = (): UseQueryResult<TAdministrationResponse> =>
