+++ src/modules/Administration/store.ts
@@ -16,3 +16,8 @@ export const AdministrationStatusState = atom({
   key: "administration-status-state",
   default: "none",
 });
+export const AdministrationStepState = atom({
+  key: "administration-step-state",
+  default: 0,
+});
+
