+++ src/modules/Administration/type.d.ts
@@ -1,3 +1,5 @@
+import { TMetaResponseSingle } from "../../services/types";
+
 export type ReturnTypesPrivateInformation = {
     setPrivateStatus: (val: boolean) => void;
     getPrivateStatus: boolean;
@@ -17,4 +19,17 @@ export type StatusReturnTypesAdministration = {
     setAdministrationStatus: (val: string) => void;
     getAdministrationStatus: string;
   };
-  
+
+export type TAdministrationItem = {
+  id: string;
+  user_id: string;
+  status: string;
+  type: string;
+  action_by: string;
+  biodata: string;
+  familial: string;
+  file: string;
+};
+
+
+export type TAdministrationResponse = TMetaResponseSingle<TAdministrationItem>;
