--- src/stores/Administration/index.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const AdministrationStepState = atom({
-  key: "administration-step-state",
-  default: 0,
-});
